timer gneauer

+++ include/Timer.h
@@ -31,9 +31,9 @@ struct Time {
 Time convertFromIndustrialHours(double industrialHours);
 void Timer_Main();
 
-extern double Timer_1;
-extern double Timer_2;
-extern double Timer_3;
+extern long double Timer_1;
+extern long double Timer_2;
+extern long double Timer_3;
 
 extern unsigned long previousMillis; // Variable zum Speichern der letzten Zeitmessung
 extern const unsigned long interval; // Intervall in Millisekunden (hier 1 Sekunde)
